+++ example/lib/main.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart' show env;
 import 'package:giphy_get/giphy_get.dart';
 import 'package:giphy_get_demo/providers/theme_provider.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart' as DotEnv;
 
 Future<void> main() async {
-  await DotEnv().load('.env');
+  WidgetsFlutterBinding.ensureInitialized();
+  await DotEnv.load(fileName: ".env");
   runApp(MultiProvider(providers: [
     ChangeNotifierProvider(
         create: (ctx) => ThemeProvider(currentTheme: ThemeMode.system))
@@ -53,19 +55,19 @@ class _MyHomePageState extends State<MyHomePage> {
   // Random ID
   String randomId = "";
 
-  String giphy_api_key = DotEnv().env["giphy_api_key"];
+  String giphy_api_key = env["giphy_api_key"];
 
   @override
   void initState() {
     super.initState();
 
-    client = GiphyClient(apiKey: giphy_api_key);
-    initAsync();
-  }
-
-  initAsync() {
-    setState(() async {
-      randomId = await client.getRandomId();
+    client = GiphyClient(apiKey: giphy_api_key, randomId: '');
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      client.getRandomId().then((value) {
+        setState(() {
+          randomId = value;
+        });
+      });
     });
   }