+++ lib/src/client/models/images.dart
@@ -1,35 +1,35 @@
 import 'package:giphy_get/src/client/models/image.dart';
 
 class GiphyImages {
-  final GiphyStillImage fixedHeightStill;
-  final GiphyStillImage originalStill;
-  final GiphyFullImage fixedWidth;
-  final GiphyStillImage fixedHeightSmallStill;
-  final GiphyDownsampledImage fixedHeightDownsampled;
-  final GiphyPreviewImage preview;
-  final GiphyFullImage fixedHeightSmall;
-  final GiphyStillImage downsizedStill;
-  final GiphyDownsizedImage downsized;
-  final GiphyDownsizedImage downsizedLarge;
-  final GiphyStillImage fixedWidthSmallStill;
-  final GiphyWebPImage previewWebp;
-  final GiphyStillImage fixedWidthStill;
-  final GiphyFullImage fixedWidthSmall;
-  final GiphyPreviewImage downsizedSmall;
-  final GiphyDownsampledImage fixedWidthDownsampled;
-  final GiphyPreviewImage downsizedMedium;
-  final GiphyOriginalImage original;
-  final GiphyFullImage fixedHeight;
-  final GiphyPreviewImage hd;
-  final GiphyLoopingImage looping;
-  final GiphyPreviewImage originalMp4;
-  final GiphyDownsizedImage previewGif;
-  final GiphyStillImage w480Still;
+  GiphyStillImage fixedHeightStill;
+  GiphyStillImage originalStill;
+  GiphyFullImage fixedWidth;
+  GiphyStillImage? fixedHeightSmallStill;
+  GiphyDownsampledImage? fixedHeightDownsampled;
+  GiphyPreviewImage? preview;
+  GiphyFullImage? fixedHeightSmall;
+  GiphyStillImage? downsizedStill;
+  GiphyDownsizedImage? downsized;
+  GiphyDownsizedImage? downsizedLarge;
+  GiphyStillImage? fixedWidthSmallStill;
+  GiphyWebPImage? previewWebp;
+  GiphyStillImage? fixedWidthStill;
+  GiphyFullImage? fixedWidthSmall;
+  GiphyPreviewImage? downsizedSmall;
+  GiphyDownsampledImage? fixedWidthDownsampled;
+  GiphyPreviewImage? downsizedMedium;
+  GiphyOriginalImage? original;
+  GiphyFullImage? fixedHeight;
+  GiphyPreviewImage? hd;
+  GiphyLoopingImage? looping;
+  GiphyPreviewImage? originalMp4;
+  GiphyDownsizedImage? previewGif;
+  GiphyStillImage? w480Still;
 
   GiphyImages({
-    this.fixedHeightStill,
-    this.originalStill,
-    this.fixedWidth,
+    required this.fixedHeightStill,
+    required this.originalStill,
+    required this.fixedWidth,
     this.fixedHeightSmallStill,
     this.fixedHeightDownsampled,
     this.preview,
@@ -53,109 +53,100 @@ class GiphyImages {
     this.w480Still,
   });
 
-  factory GiphyImages.fromJson(Map<String, dynamic> json) {
+  factory GiphyImages.fromJson(Map<String, dynamic?> json) {
     return GiphyImages(
-      fixedHeightStill: json['fixed_height_still'] == null
+      fixedHeightStill: GiphyStillImage.fromJson(json['fixed_height_still']),
+      originalStill: GiphyStillImage.fromJson(json['original_still']),
+      fixedWidth: GiphyFullImage.fromJson(json['fixed_width']),
+      fixedHeightSmallStill: json['fixed_height_small_still'] == null ||
+              (json['fixed_height_small_still'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyStillImage.fromJson(
-              json['fixed_height_still'] as Map<String, dynamic>),
-      originalStill: json['original_still'] == null
+          : GiphyStillImage.fromJson(json['fixed_height_small_still']),
+      fixedHeightDownsampled: json['fixed_height_downsampled'] == null ||
+              (json['fixed_height_downsampled'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyStillImage.fromJson(
-              json['original_still'] as Map<String, dynamic>),
-      fixedWidth: json['fixed_width'] == null
+          : GiphyDownsampledImage.fromJson(json['fixed_height_downsampled']),
+      preview: json['preview'] == null ||
+              (json['preview'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyFullImage.fromJson(
-              json['fixed_width'] as Map<String, dynamic>),
-      fixedHeightSmallStill: json['fixed_height_small_still'] == null
+          : GiphyPreviewImage.fromJson(json['preview']),
+      fixedHeightSmall: json['fixed_height_small'] == null ||
+              (json['fixed_height_small'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyStillImage.fromJson(
-              json['fixed_height_small_still'] as Map<String, dynamic>),
-      fixedHeightDownsampled: json['fixed_height_downsampled'] == null
+          : GiphyFullImage.fromJson(json['fixed_height_small']),
+      downsizedStill: json['downsized_still'] == null ||
+              (json['downsized_still'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyDownsampledImage.fromJson(
-              json['fixed_height_downsampled'] as Map<String, dynamic>),
-      preview: json['preview'] == null
+          : GiphyStillImage.fromJson(json['downsized_still']),
+      downsized: json['downsized'] == null ||
+              (json['downsized'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyPreviewImage.fromJson(json['preview'] as Map<String, dynamic>),
-      fixedHeightSmall: json['fixed_height_small'] == null
+          : GiphyDownsizedImage.fromJson(json['downsized']),
+      downsizedLarge: json['downsized_large'] == null ||
+              (json['downsized_large'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyFullImage.fromJson(
-              json['fixed_height_small'] as Map<String, dynamic>),
-      downsizedStill: json['downsized_still'] == null
+          : GiphyDownsizedImage.fromJson(json['downsized_large']),
+      fixedWidthSmallStill: json['fixed_width_small_still'] == null ||
+              (json['fixed_width_small_still'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyStillImage.fromJson(
-              json['downsized_still'] as Map<String, dynamic>),
-      downsized: json['downsized'] == null
+          : GiphyStillImage.fromJson(json['fixed_width_small_still']),
+      previewWebp: json['preview_webp'] == null ||
+              (json['preview_webp'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyDownsizedImage.fromJson(
-              json['downsized'] as Map<String, dynamic>),
-      downsizedLarge: json['downsized_large'] == null
+          : GiphyWebPImage.fromJson(json['preview_webp']),
+      fixedWidthStill: json['fixed_width_still'] == null ||
+              (json['fixed_width_still'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyDownsizedImage.fromJson(
-              json['downsized_large'] as Map<String, dynamic>),
-      fixedWidthSmallStill: json['fixed_width_small_still'] == null
+          : GiphyStillImage.fromJson(json['fixed_width_still']),
+      fixedWidthSmall: json['fixed_width_small'] == null ||
+              (json['fixed_width_small'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyStillImage.fromJson(
-              json['fixed_width_small_still'] as Map<String, dynamic>),
-      previewWebp: json['preview_webp'] == null
+          : GiphyFullImage.fromJson(json['fixed_width_small']),
+      downsizedSmall: json['downsized_small'] == null ||
+              (json['downsized_small'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyWebPImage.fromJson(
-              json['preview_webp'] as Map<String, dynamic>),
-      fixedWidthStill: json['fixed_width_still'] == null
+          : GiphyPreviewImage.fromJson(json['downsized_small']),
+      fixedWidthDownsampled: json['fixed_width_downsampled'] == null ||
+              (json['fixed_width_downsampled'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyStillImage.fromJson(
-              json['fixed_width_still'] as Map<String, dynamic>),
-      fixedWidthSmall: json['fixed_width_small'] == null
+          : GiphyDownsampledImage.fromJson(json['fixed_width_downsampled']),
+      downsizedMedium: json['downsized_medium'] == null ||
+              (json['downsized_medium'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyFullImage.fromJson(
-              json['fixed_width_small'] as Map<String, dynamic>),
-      downsizedSmall: json['downsized_small'] == null
+          : GiphyPreviewImage.fromJson(json['downsized_medium']),
+      original: json['original'] == null ||
+              (json['original'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyPreviewImage.fromJson(
-              json['downsized_small'] as Map<String, dynamic>),
-      fixedWidthDownsampled: json['fixed_width_downsampled'] == null
+          : GiphyOriginalImage.fromJson(json['original']),
+      fixedHeight: json['fixed_height'] == null ||
+              (json['fixed_height'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyDownsampledImage.fromJson(
-              json['fixed_width_downsampled'] as Map<String, dynamic>),
-      downsizedMedium: json['downsized_medium'] == null
+          : GiphyFullImage.fromJson(json['fixed_height']),
+      hd: json['hd'] == null ? null : GiphyPreviewImage.fromJson(json['hd']),
+      looping: json['looping'] == null ||
+              (json['looping'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyPreviewImage.fromJson(
-              json['downsized_medium'] as Map<String, dynamic>),
-      original: json['original'] == null
+          : GiphyLoopingImage.fromJson(json['looping']),
+      originalMp4: json['original_mp4'] == null ||
+              (json['original_mp4'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyOriginalImage.fromJson(
-              json['original'] as Map<String, dynamic>),
-      fixedHeight: json['fixed_height'] == null
+          : GiphyPreviewImage.fromJson(json['original_mp4']),
+      previewGif: json['preview_gif'] == null ||
+              (json['preview_gif'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyFullImage.fromJson(
-              json['fixed_height'] as Map<String, dynamic>),
-      hd: json['hd'] == null
+          : GiphyDownsizedImage.fromJson(json['preview_gif']),
+      w480Still: json['480w_still'] == null ||
+              (json['480w_still'] as Map<String, dynamic>).isEmpty
           ? null
-          : GiphyPreviewImage.fromJson(json['hd'] as Map<String, dynamic>),
-      looping: json['looping'] == null
-          ? null
-          : GiphyLoopingImage.fromJson(json['looping'] as Map<String, dynamic>),
-      originalMp4: json['original_mp4'] == null
-          ? null
-          : GiphyPreviewImage.fromJson(
-              json['original_mp4'] as Map<String, dynamic>),
-      previewGif: json['preview_gif'] == null
-          ? null
-          : GiphyDownsizedImage.fromJson(
-              json['preview_gif'] as Map<String, dynamic>),
-      w480Still: json['480w_still'] == null
-          ? null
-          : GiphyStillImage.fromJson(
-              json['480w_still'] as Map<String, dynamic>),
+          : GiphyStillImage.fromJson(json['480w_still']),
     );
   }
 
   Map<String, dynamic> toJson() {
     return <String, dynamic>{
-      'fixed_height_still': fixedHeightStill?.toJson(),
-      'original_still': originalStill?.toJson(),
-      'fixed_width': fixedWidth?.toJson(),
+      'fixed_height_still': fixedHeightStill.toJson(),
+      'original_still': originalStill.toJson(),
+      'fixed_width': fixedWidth.toJson(),
       'fixed_height_small_still': fixedHeightSmallStill?.toJson(),
       'fixed_height_downsampled': fixedHeightDownsampled?.toJson(),
       'preview': preview?.toJson(),