+++ lib/src/views/appbar/searchappbar.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:giphy_get/src/client/models/type.dart';
 import 'package:giphy_get/src/providers/app_bar_provider.dart';
@@ -9,7 +10,7 @@ class SearchAppBar extends StatefulWidget {
   // Scroll Controller
   final ScrollController scrollController;
 
-  SearchAppBar({Key key, this.scrollController}) : super(key: key);
+  SearchAppBar({Key? key, required this.scrollController}) : super(key: key);
 
   @override
   _SearchAppBarState createState() => _SearchAppBarState();
@@ -17,13 +18,13 @@ class SearchAppBar extends StatefulWidget {
 
 class _SearchAppBarState extends State<SearchAppBar> {
   // Tab Provider
-  TabProvider _tabProvider;
+  late TabProvider _tabProvider;
 
   // AppBar Provider
-  AppBarProvider _appBarProvider;
+  late AppBarProvider _appBarProvider;
 
   // Sheet Provider
-  SheetProvider _sheetProvider;
+  late SheetProvider _sheetProvider;
 
   // Input controller
   final TextEditingController _textEditingController =
@@ -32,11 +33,11 @@ class _SearchAppBarState extends State<SearchAppBar> {
   final FocusNode _focus = new FocusNode();
 
   //Colors
-  Color _canvasColor;
-  Color _searchBackgroundColor;
+  late Color _canvasColor;
+  late Color _searchBackgroundColor;
 
   //Is DarkMode
-  bool _isDarkMode;
+  late bool _isDarkMode;
 
   @override
   void initState() {
@@ -57,7 +58,7 @@ class _SearchAppBarState extends State<SearchAppBar> {
   void didChangeDependencies() {
     //Colors
     _canvasColor = Theme.of(context).canvasColor;
-    _searchBackgroundColor = Theme.of(context).textTheme.bodyText1.color;
+    _searchBackgroundColor = Theme.of(context).textTheme.bodyText1!.color!;
 
     //Is DarkMode
     _isDarkMode = Theme.of(context).brightness == Brightness.dark;
@@ -136,20 +137,27 @@ class _SearchAppBarState extends State<SearchAppBar> {
     return Center(
         child: Image.asset(
       "$basePath$logoPath",
+      width: 100.0,
       package: 'giphy_get',
     ));
   }
 
-  Widget _searchIcon() => ShaderMask(
+  Widget _searchIcon() {
+    if (kIsWeb) {
+      return Icon(Icons.search);
+    } else {
+      return ShaderMask(
         shaderCallback: (bounds) => LinearGradient(
             begin: Alignment.topLeft,
             end: Alignment.bottomRight,
             colors: [
               Colors.pinkAccent,
-              Colors.purple[700],
+              Colors.purple[700]!,
             ]).createShader(bounds),
         child: Icon(Icons.search),
       );
+    }
+  }
 
   void _focusListener() {
     // Set to max extent height if Textfield has focus