+++ lib/src/views/tab/giphy_tab_detail.dart
@@ -15,7 +15,7 @@ import 'package:provider/provider.dart';
 class GiphyTabDetail extends StatefulWidget {
   final String type;
   final ScrollController scrollController;
-  GiphyTabDetail({Key key, @required this.type, this.scrollController})
+  GiphyTabDetail({Key? key, required this.type, required this.scrollController})
       : super(key: key);
 
   @override
@@ -24,31 +24,31 @@ class GiphyTabDetail extends StatefulWidget {
 
 class _GiphyTabDetailState extends State<GiphyTabDetail> {
   // Tab Provider
-  TabProvider _tabProvider;
+  late TabProvider _tabProvider;
 
   // AppBar Provider
-  AppBarProvider _appBarProvider;
+  late AppBarProvider _appBarProvider;
 
   // Collection
-  GiphyCollection _collection;
+  GiphyCollection? _collection;
 
   // List of gifs
   List<GiphyGif> _list = [];
 
   // Direction
-  Axis _scrollDirection;
+  final Axis _scrollDirection = Axis.vertical;
 
   // Axis count
-  int _crossAxisCount;
+  late int _crossAxisCount;
 
   // Spacing between gifs in grid
   double _spacing = 8.0;
 
   // Default gif with
-  double _gifWidth;
+  late double _gifWidth;
 
   // Limit of query
-  int _limit;
+  late int _limit;
 
   // is Loading gifs
   bool _isLoading = false;
@@ -62,9 +62,6 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
 
     // AppBar Provider
     _appBarProvider = Provider.of<AppBarProvider>(context, listen: false);
-
-    // GridOptions
-    _scrollDirection = Axis.vertical;
 
     // Gif WIDTH
     switch (widget.type) {
@@ -80,17 +77,17 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
       default:
         break;
     }
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
 
     // ScrollController
     widget.scrollController..addListener(_scrollListener);
 
     // Listen query
     _appBarProvider.addListener(_listenerQuery);
-  }
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
 
     // Set items count responsive
     _crossAxisCount = (MediaQuery.of(context).size.width / _gifWidth).round();
@@ -133,8 +130,6 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
           crossAxisSpacing: _spacing,
           itemBuilder: (ctx, idx) {
             GiphyGif _gif = _list[idx];
-            if (_gif == null) return Container();
-
             return _item(_gif);
           },
           staggeredTileBuilder: (idx) => StaggeredTile.fit(1)),
@@ -146,58 +141,61 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
         borderRadius: BorderRadius.circular(10.0),
         child: InkWell(
             onTap: () => _selectedGif(gif),
-            child: ExtendedImage.network(gif.images.fixedWidth.webp,
-                cache: true,
-                fit: BoxFit.fill,
-                headers: {'accept': 'image/*'}, loadStateChanged: (state) {
-              return AnimatedSwitcher(
-                duration: const Duration(milliseconds: 350),
-                child: case2(
-                    state.extendedImageLoadState,
-                    {
-                      LoadState.loading: Container(
-                        color: Theme.of(context).cardColor,
-                        width: _gifWidth,
-                        height: double.parse(gif.images.fixedWidth.height) *
-                            (_gifWidth /
-                                double.parse(gif.images.fixedWidth.width)),
-                      ),
-                      LoadState.completed: Menu(
-                        child: ExtendedRawImage(
-                          image: state.extendedImageInfo?.image,
-                          width: _gifWidth,
-                          height: double.parse(gif.images.fixedWidth.height) *
-                              (_gifWidth /
-                                  double.parse(gif.images.fixedWidth.width)),
-                          fit: widget.type == GiphyType.gifs
-                              ? BoxFit.fill
-                              : BoxFit.fitWidth,
-                        ),
-                        items: [
-                          if (gif.username != null && gif.username.isNotEmpty)
-                            MenuItem("More @${gif.username}", () {
-                              _appBarProvider.queryText = gif.username;
-                            }),
-                        ],
-                        decoration: MenuDecoration(),
-                      ),
-                      LoadState.failed: Container(
-                        color: Theme.of(context).cardColor,
-                        width: _gifWidth,
-                        height: double.parse(gif.images.fixedWidth.height) *
-                            (_gifWidth /
-                                double.parse(gif.images.fixedWidth.width)),
-                      ),
-                    },
-                    Container(
-                      color: Theme.of(context).cardColor,
-                      width: _gifWidth,
-                      height: double.parse(gif.images.fixedWidth.height) *
-                          (_gifWidth /
-                              double.parse(gif.images.fixedWidth.width)),
-                    )),
-              );
-            })));
+            child: gif.images == null || gif.images?.fixedWidth.webp == null
+                ? Container()
+                : ExtendedImage.network(gif.images!.fixedWidth.webp!,
+                    cache: true,
+                    fit: BoxFit.fill,
+                    headers: {'accept': 'image/*'}, loadStateChanged: (state) {
+                    return AnimatedSwitcher(
+                      duration: const Duration(milliseconds: 350),
+                      child: gif.images == null
+                          ? Container()
+                          : case2(
+                              state.extendedImageLoadState,
+                              {
+                                LoadState.loading: Container(
+                                  color: Theme.of(context).cardColor,
+                                  width: _gifWidth,
+                                  height: double.parse(
+                                          gif.images!.fixedWidth.height) *
+                                      (_gifWidth /
+                                          double.parse(
+                                              gif.images!.fixedWidth.width)),
+                                ),
+                                LoadState.completed: ExtendedRawImage(
+                                  image: state.extendedImageInfo?.image,
+                                  width: _gifWidth,
+                                  height: double.parse(
+                                          gif.images!.fixedWidth.height) *
+                                      (_gifWidth /
+                                          double.parse(
+                                              gif.images!.fixedWidth.width)),
+                                  fit: widget.type == GiphyType.gifs
+                                      ? BoxFit.fill
+                                      : BoxFit.fitWidth,
+                                ),
+                                LoadState.failed: Container(
+                                  color: Theme.of(context).cardColor,
+                                  width: _gifWidth,
+                                  height: double.parse(
+                                          gif.images!.fixedWidth.height) *
+                                      (_gifWidth /
+                                          double.parse(
+                                              gif.images!.fixedWidth.width)),
+                                ),
+                              },
+                              Container(
+                                color: Theme.of(context).cardColor,
+                                width: _gifWidth,
+                                height: double.parse(
+                                        gif.images!.fixedWidth.height) *
+                                    (_gifWidth /
+                                        double.parse(
+                                            gif.images!.fixedWidth.width)),
+                              )),
+                    );
+                  })));
   }
 
   Future<void> _loadMore() async {
@@ -212,9 +210,12 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
         apiKey: _tabProvider.apiKey, randomId: _tabProvider.randomID);
 
     // Offset pagination for query
-    int offset = _collection == null
-        ? 0
-        : _collection.pagination.offset + _collection.pagination.count;
+    int offset;
+    if (_collection == null) {
+      offset = 0;
+    } else {
+      offset = _collection!.pagination!.offset + _collection!.pagination!.count;
+    }
 
     // Get Gif or Emoji
     if (widget.type == GiphyType.emoji) {
@@ -240,9 +241,9 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
     }
 
     // Set result to list
-    if (_collection.data.isNotEmpty && mounted) {
+    if (_collection!.data!.isNotEmpty && mounted) {
       setState(() {
-        _list.addAll(_collection.data);
+        _list.addAll(_collection!.data!);
       });
     }
 
@@ -273,10 +274,10 @@ class _GiphyTabDetailState extends State<GiphyTabDetail> {
     _loadMore();
   }
 
-  TValue case2<TOptionType, TValue>(
+  TValue? case2<TOptionType, TValue>(
     TOptionType selectedOption,
     Map<TOptionType, TValue> branches, [
-    TValue defaultValue = null,
+    TValue? defaultValue = null,
   ]) {
     if (!branches.containsKey(selectedOption)) {
       return defaultValue;